chore: 폴더 위치 변경 및 storage 로직 위치 변경
