+++ src/pages/popup/Popup.tsx
@@ -1,6 +1,6 @@
 import '@pages/popup/Popup.css';
 import withSuspense from '@src/shared/hoc/withSuspense';
-import { Tags } from '@pages/popup/Tag';
+import { Tags } from '@pages/popup/components/Tag';
 import styled from 'styled-components';
 
 const Popup = () => {
